requirements.txt
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
openai
streamlit
crewai==0.5.5
